New public API
