--- src/NATS.Client.Core/NatsConnection.Command.cs
@@ -1,60 +1,0 @@
-using System.Runtime.CompilerServices;
-using Microsoft.Extensions.Logging;
-using NATS.Client.Core.Commands;
-using NATS.Client.Core.Internal;
-
-namespace NATS.Client.Core;
-
-public partial class NatsConnection : INatsCommand
-{
-    public void PostPing(CancellationToken cancellationToken = default)
-    {
-        if (ConnectionState == NatsConnectionState.Open)
-        {
-            EnqueueCommandSync(PingCommand.Create(_pool, GetCommandTimer(cancellationToken)));
-        }
-        else
-        {
-            WithConnect(cancellationToken, static (self, token) => self.EnqueueCommandSync(PingCommand.Create(self._pool, self.GetCommandTimer(token))));
-        }
-    }
-
-    // public ValueTask PostPingAsync(CancellationToken cancellationToken = default)
-    // {
-    //     if (ConnectionState == NatsConnectionState.Open)
-    //     {
-    //         return EnqueueCommandAsync(PingCommand.Create(_pool, GetCommandTimer(cancellationToken)));
-    //     }
-    //     else
-    //     {
-    //         return WithConnectAsync(cancellationToken, static (self, token) => self.EnqueueCommandAsync(PingCommand.Create(self._pool, self.GetCommandTimer(token))));
-    //     }
-    // }
-
-    /// <summary>
-    /// Send PING command and await PONG. Return value is similar as Round trip time.
-    /// </summary>
-    public ValueTask<TimeSpan> PingAsync(CancellationToken cancellationToken = default)
-    {
-        if (ConnectionState == NatsConnectionState.Open)
-        {
-            var command = AsyncPingCommand.Create(this, _pool, GetCommandTimer(cancellationToken));
-            if (TryEnqueueCommand(command))
-            {
-                return command.AsValueTask();
-            }
-            else
-            {
-                return EnqueueAndAwaitCommandAsync(command);
-            }
-        }
-        else
-        {
-            return WithConnectAsync(cancellationToken, static (self, token) =>
-            {
-                var command = AsyncPingCommand.Create(self, self._pool, self.GetCommandTimer(token));
-                return self.EnqueueAndAwaitCommandAsync(command);
-            });
-        }
-    }
-}
