+++ src/NATS.Client.Core/NatsPubOpts.cs
@@ -4,7 +4,11 @@ public readonly record struct NatsPubOpts
 {
     public string? ReplyTo { get; init; }
 
-    public NatsHeaders? Headers { get; init; }
-
+    // TODO: Implement headers in NatsPubOpts
+    // public NatsHeaders? Headers
+    // {
+    //     get => throw new NotImplementedException();
+    //     init => throw new NotImplementedException();
+    // }
     public INatsSerializer? Serializer { get; init; }
 }
